Car Payment
Create an application with 1 form.  On this form I want you to calculate
a monthly car payment for buying a new car.  You need to take in the
following as user input:  Name of the vehicle (Ford Mustang for
example), Purchase Price, Sales Tax Amount, and Number of Years for the
loan.

You will need to have a button to calculate the result, clear the user
input, and exit the application.  When the user hits the calculate
button you need to display the car name and the result of the car
payment in read-only text boxes or a labels (your choice).

When the user hits the Clear button, you need to clear the user’s entry
out of the text boxes.  When the user hits the exit button, the
application needs to close.

Please use variables with the correct data type to capture the user’s
input.  To simplify the calculation, assume no annual interest rate and
use this formula: Car Payment = (Purchase Price * Sales Tax) / (Years *
12).

Please also include a Status Strip control at the bottom of the form
with a Status Label that displays the current date and time when the
form loads (Use the Forms On-Load event).

How you design the form and name the controls and variables is up to,
but make the layout make sense.